move last remaining auth helpers to dashboard
