+++ pegasus/helpers.rb
@@ -1,7 +1,6 @@
 require 'cdo/aws/s3'
 require 'rack/csrf'
 require_relative '../shared/middleware/helpers/core.rb'
-require_relative '../shared/middleware/helpers/shared_auth_helpers.rb'
 require 'cdo/asset_helper'
 require 'cdo/cookie_helpers'
 require 'cdo/language_dir'
