--- shared/middleware/helpers/shared_auth_helpers.rb
@@ -1,22 +1,0 @@
-require 'cdo/user_helpers'
-require 'json'
-
-#
-# Utility methods that help middleware access dashboard authentication and
-# permissions information.
-#
-# The use of these methods within shared and pegasus is deprecated.
-# These methods should be eliminated or moved to dashboard.
-#
-
-# @returns [Integer] the user_id associated with the current request
-def current_user_id
-  # @request is a Sinatra::Request < Rack::Request provided by Sinatra::Base
-  @request.nil? ? nil : @request.user_id
-end
-
-# @returns [Hash] the dashboard user row associated with the current request.
-def current_user
-  return nil if (id = current_user_id).nil?
-  @dashboard_user ||= DASHBOARD_DB[:users][id: id]
-end
